[monaco] update model options once on a preference change, not for each editor
